+++ packages/expo-router/src/layouts/stack-utils/toolbar/bottom-toolbar-native-elements.tsx
@@ -5,7 +5,6 @@ import {
   Platform,
   StyleSheet,
   type ColorValue,
-  type ImageSourcePropType,
   type StyleProp,
   type TextStyle,
 } from 'react-native';
@@ -13,9 +12,7 @@ import type { SFSymbol } from 'sf-symbols-typescript';
 
 import { LinkMenuAction } from '../../../link/elements';
 import { NativeLinkPreviewAction } from '../../../link/preview/native';
-import { Label, Icon } from '../../../primitives';
 import { RouterToolbarItem } from '../../../toolbar/native';
-import { getFirstChildOfType } from '../../../utils/children';
 import type { BasicTextStyle } from '../../../utils/font';
 
 // #region NativeToolbarMenu
@@ -29,7 +26,7 @@ export interface NativeToolbarMenuProps {
   disabled?: boolean;
   hidden?: boolean;
   hidesSharedBackground?: boolean;
-  icon?: SFSymbol | ImageSourcePropType;
+  icon?: SFSymbol;
   // TODO(@ubax): Add useImage support in a follow-up PR.
   /**
    * Image to display for the menu item.
@@ -73,12 +70,7 @@ export const NativeToolbarMenu: React.FC<NativeToolbarMenuProps> = ({
   elementSize,
 }) => {
   const identifier = useId();
-  const iconComponent = getFirstChildOfType(children, Icon);
 
-  const computedIcon =
-    icon ??
-    (iconComponent?.props && 'sf' in iconComponent.props ? iconComponent.props.sf : undefined);
-  const sf = typeof computedIcon === 'string' ? computedIcon : undefined;
   const titleStyle = StyleSheet.flatten(style);
   const renderingMode = imageRenderingMode ?? (tintColor !== undefined ? 'template' : 'original');
   return (
@@ -86,7 +78,7 @@ export const NativeToolbarMenu: React.FC<NativeToolbarMenuProps> = ({
       sharesBackground={!separateBackground}
       hidesSharedBackground={hidesSharedBackground}
       hidden={hidden}
-      icon={sf}
+      icon={icon}
       // TODO(@ubax): Handle image loading using useImage in a follow-up PR.
       image={image}
       imageRenderingMode={renderingMode}
@@ -125,7 +117,6 @@ export const NativeToolbarMenuAction = LinkMenuAction;
 export interface NativeToolbarButtonProps {
   accessibilityLabel?: string;
   accessibilityHint?: string;
-  children?: ReactNode;
   disabled?: boolean;
   hidden?: boolean;
   hidesSharedBackground?: boolean;
@@ -139,6 +130,7 @@ export interface NativeToolbarButtonProps {
   style?: StyleProp<BasicTextStyle>;
   tintColor?: ColorValue;
   variant?: 'plain' | 'done' | 'prominent';
+  label?: string;
 }
 
 /**
@@ -147,16 +139,6 @@ export interface NativeToolbarButtonProps {
  */
 export const NativeToolbarButton: React.FC<NativeToolbarButtonProps> = (props) => {
   const id = useId();
-  const areChildrenString = typeof props.children === 'string';
-  const label = areChildrenString
-    ? (props.children as string)
-    : getFirstChildOfType(props.children, Label)?.props.children;
-  const iconComponent =
-    !props.icon && !areChildrenString ? getFirstChildOfType(props.children, Icon) : undefined;
-  const icon =
-    props.icon ??
-    (iconComponent?.props && 'sf' in iconComponent.props ? iconComponent.props.sf : undefined);
-  const sf = typeof icon === 'string' ? icon : undefined;
   const renderingMode =
     props.imageRenderingMode ?? (props.tintColor !== undefined ? 'template' : 'original');
   return (
@@ -174,8 +156,8 @@ export const NativeToolbarButton: React.FC<NativeToolbarButtonProps> = (props) =
       possibleTitles={props.possibleTitles}
       selected={props.selected}
       sharesBackground={!props.separateBackground}
-      systemImageName={sf}
-      title={label}
+      systemImageName={props.icon}
+      title={props.label}
       tintColor={props.tintColor}
       titleStyle={StyleSheet.flatten(props.style)}
     />